# src/main/resources/static/swagger.yaml
openapi: 3.0.3
info:
  title: VulnPrint API Documentation
  description: |
    ### Master Node API Documentation
    Professional endpoint registry for the VulnPrint Pentest Management Platform.
    
    **Security Notice:** All data handled by these endpoints is sensitive. 
    Ensure access is restricted to authorized security personnel.
  version: 1.1.0
servers:
  - url: http://localhost:8080
    description: Local Development Node

tags:
  - name: Authentication
    description: Identity verification and session initialization
  - name: Dashboard
    description: System-wide analytics and project summaries
  - name: Pentest
    description: Security audit project management
  - name: Vulnerability
    description: Technical finding registry and evidence management
  - name: Organization
    description: Global branding and identity configuration
  - name: User
    description: Operator registry and access control
  - name: Report
    description: Data consolidation for automated generation

paths:
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Initialize Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Access Granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Access Denied

  /api/dashboard/metrics:
    get:
      tags: [Dashboard]
      summary: Retrieve Global Analytics
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardMetrics'

  /api/dashboard/pentests:
    get:
      tags: [Dashboard]
      summary: List Project Nodes
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 3 }
        - name: type
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Enriched project list

  /api/pentests:
    post:
      tags: [Pentest]
      summary: Register/Update Project Node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pentest'
      responses:
        '200':
          description: Node Synced

  /api/pentests/{id}:
    get:
      tags: [Pentest]
      summary: Retrieve Node Attributes
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Metadata retrieved
    delete:
      tags: [Pentest]
      summary: Terminate Project Node
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Node purged

  /api/pentests/{id}/report-details:
    put:
      tags: [Pentest]
      summary: Synchronize Report Design
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Configuration updated

  /api/organization:
    get:
      tags: [Organization]
      summary: Retrieve Master Branding
      responses:
        '200':
          description: Success
    put:
      tags: [Organization]
      summary: Update Master Branding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organization'
      responses:
        '200':
          description: Branding synced

  /api/users:
    get:
      tags: [User]
      summary: List All Operators
      responses:
        '200':
          description: Operator list retrieved

  /api/users/profile/{username}:
    get:
      tags: [User]
      summary: Retrieve Operator Profile
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Profile attributes

  /api/users/profile:
    post:
      tags: [User]
      summary: Update Operator Profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success

  /api/users/register:
    post:
      tags: [User]
      summary: Initialize New Operator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Operator created

  /api/users/change-password:
    post:
      tags: [User]
      summary: Update Operator Passcode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                currentPassword: { type: string }
                newPassword: { type: string }
      responses:
        '200':
          description: Success

  /api/vulnerabilities:
    post:
      tags: [Vulnerability]
      summary: Register Security Finding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vulnerability'
      responses:
        '200':
          description: Finding initialized

  /api/vulnerabilities/{id}:
    get:
      tags: [Vulnerability]
      summary: Retrieve Finding Attributes
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Metadata retrieved
    put:
      tags: [Vulnerability]
      summary: Update Finding Data
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vulnerability'
      responses:
        '200':
          description: Metadata updated
    delete:
      tags: [Vulnerability]
      summary: Purge Finding Registry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Finding removed

  /api/vulnerabilities/poc-image:
    get:
      tags: [Vulnerability]
      summary: Retrieve POC Image Stream
      parameters:
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Image stream binary

  /api/repGenApi/{pentestId}:
    get:
      tags: [Report]
      summary: Generate Final Report Data Package
      parameters:
        - name: pentestId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Consolidated report data (V3/repGenApi format)

components:
  schemas:
    AuthResponse:
      type: object
      properties:
        status: { type: string }
        token: { type: string }
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        role: { type: string }
        profileImage: { type: string }
        message: { type: string }

    DashboardMetrics:
      type: object
      properties:
        totalPentests: { type: integer }
        totalVulnerabilities: { type: integer }
        critical: { type: integer }
        high: { type: integer }
        medium: { type: integer }
        low: { type: integer }
        info: { type: integer }
        globalRiskScore: { type: string }

    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        role: { type: string }
        address: { type: string }
        qualification: { type: string }

    Organization:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: string }
        logo: { type: string }

    Pentest:
      type: object
      properties:
        id: { type: integer }
        pentestName: { type: string }
        pentestType: { type: string }
        clientName: { type: string }
        applicationName: { type: string }
        status: { type: string }
        assignedPentesters:
          type: array
          items:
            $ref: '#/components/schemas/User'

    Vulnerability:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        severity: { type: string }
        cvssScore: { type: number }
        cvssVector: { type: string }
        description: { type: string }
        impact: { type: string }
        mitigation: { type: string }
        evidenceMode: { type: string, enum: [STEPS, REQRES, SOURCE] }
        steps:
          type: array
          items:
            $ref: '#/components/schemas/VulnerabilityStep'
        pentest:
          type: object
          properties:
            id: { type: integer }

    VulnerabilityStep:
      type: object
      properties:
        stepNumber: { type: integer }
        description: { type: string }
        request: { type: string }
        response: { type: string }
        fileName: { type: string }
        lineNumber: { type: string }
        imagePaths:
          type: array
          items: { type: string }